+++ Runtime/CDT/CDT.Primitive.cs
@@ -25,12 +25,12 @@ namespace Voxell.GPUVectorGraphics
         (this.e0 == other.e1 && this.e1 == other.e0);
     }
 
-    private struct Circumcenter
+    private struct Cirumcircle
     {
       public float2 center;
       public float sqradius;
 
-      public Circumcenter(float2 p0, float2 p1, float2 p2)
+      public Cirumcircle(float2 p0, float2 p1, float2 p2)
       {
         float dA = p0.x * p0.x + p0.y * p0.y;
         float dB = p1.x * p1.x + p1.y * p1.y;
@@ -39,7 +39,6 @@ namespace Voxell.GPUVectorGraphics
         float aux1 = dA * (p2.y - p1.y) + dB * (p0.y - p2.y) + dC * (p1.y - p0.y);
         float aux2 = -(dA * (p2.x - p1.x) + dB * (p0.x - p2.x) + dC * (p1.x - p0.x));
         float div = 2.0f * (p0.x * (p2.y - p1.y) + p1.x * (p0.y - p2.y) + p2.x * (p1.y - p0.y));
-        div += math.EPSILON;
         div = 1.0f / div;
 
         center = new float2(aux1, aux2) * div;